requirements.txt
flask==3.0.0
gunicorn==22.0.0
alembic==1.14.1
requests==2.31.0
python-dateutil==2.8.2
python-dotenv==1.0.1
SQLAlchemy==2.0.38
PyMySQL==1.1.1
